The mock balance should be a u64

+++ pallets/subtensor/tests/mock.rs
@@ -44,7 +44,7 @@ pub type AccountId = u64;
 
 // Balance of an account.
 #[allow(dead_code)]
-pub type Balance = u128;
+pub type Balance = u64;
 
 // An index to a block.
 #[allow(dead_code)]
